feat(pipeline): extract static analysis and add multi-template support (#305)
