+++ autograder/steps/load_template_step.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Union, List, Optional
 
 from autograder.models.dataclass.step_result import StepResult, StepName, StepStatus
 from autograder.models.pipeline_execution import PipelineExecution
@@ -10,15 +11,17 @@ logger = logging.getLogger(__name__)
 
 class TemplateLoaderStep(Step):
     """
-    Step that loads a grading template, which contains test functions and helper code used for grading.
-    It can load either a built-in template from the library or a custom template provided by the user.
-    If the template is custom, it should be loaded in a sandboxed environment to ensure security and isolation.
+    Step that loads one or more grading templates, which contain test functions 
+    and helper code used for grading.
     """
-    def __init__(self, template_name: str, custom_template = None):
+    def __init__(self, template_name: Union[str, List[str]], custom_template = None):
         """
         Initialize the template loader step.
         """
-        self._template_name = template_name
+        if isinstance(template_name, str):
+            self._template_names = [name.strip() for name in template_name.split(",")]
+        else:
+            self._template_names = template_name
         self._custom_template = custom_template
         self._template_service = TemplateLibraryService.get_instance()
 
@@ -28,31 +31,33 @@ class TemplateLoaderStep(Step):
 
     def _execute(self, pipeline_exec: PipelineExecution) -> PipelineExecution:
         """
-        Load the grading template, either built-in or custom, and return it as part of the step result.
+        Load the grading templates and return them as part of the step result.
         """
+        templates = []
+        
         if self._custom_template:
             logger.info("Loading custom template (external_user_id=%s)", pipeline_exec.submission.user_id)
-            template = self._template_service.load_custom_template(self._custom_template) #TODO: Implement Custom Template Loading with Sandboxed Env
+            template = self._template_service.load_custom_template(self._custom_template)
+            templates.append(template)
         else:
-            logger.info(
-                "Loading built-in template: template=%s (external_user_id=%s)",
-                self._template_name,
-                pipeline_exec.submission.user_id,
-            )
-            template = self._template_service.load_builtin_template(self._template_name) # Load built-in template similar to custom to avoid code duplication
+            for name in self._template_names:
+                logger.info(
+                    "Loading built-in template: template=%s (external_user_id=%s)",
+                    name,
+                    pipeline_exec.submission.user_id,
+                )
+                template = self._template_service.load_builtin_template(name)
+                templates.append(template)
         
         logger.info(
-            "Template loaded successfully: template=%s (external_user_id=%s)",
-            self._template_name,
+            "Templates loaded successfully: count=%d (external_user_id=%s)",
+            len(templates),
             pipeline_exec.submission.user_id,
         )
         return pipeline_exec.add_step_result(
             StepResult(
                 step=StepName.LOAD_TEMPLATE,
-                data=template,
+                data=templates,
                 status=StepStatus.SUCCESS
             )
         )
-
-
-
