+++ autograder/template_library/input_output.py
@@ -357,344 +357,6 @@ class ExpectFileArtifactTest(BaseExecutionTest):
                                    path=path, expected=expected, actual=actual))
 
 
-# ===============================================================
-# TestFunction for Forbidden Import Detection
-# ===============================================================
-
-class ForbiddenImportTest(TestFunction):
-    """
-    Tests that a submission does NOT import any of the specified forbidden libraries.
-
-    Performs static analysis on submission file contents using language-aware
-    regex patterns. Supports Python, Java, JavaScript/Node, C and C++.
-    """
-
-    # Language-specific regex builders: each returns a compiled pattern
-    # that matches an import of the given library name.
-    IMPORT_PATTERNS = {
-        Language.PYTHON: [
-            # import lib  /  import lib as x  /  import lib.sub
-            r'^\s*import\s+{lib}\b',
-            # from lib import ...  /  from lib.sub import ...
-            r'^\s*from\s+{lib}\b',
-        ],
-        Language.JAVA: [
-            # import pkg.Class;  /  import static pkg.Class.method;
-            r'^\s*import\s+(?:static\s+)?{lib}\b',
-        ],
-        Language.NODE: [
-            # require('lib')  /  require("lib")
-            r"\brequire\s*\(\s*['\"]{{lib}}['\"]\s*\)",
-            # import ... from 'lib'  /  import 'lib'
-            r'^\s*import\s+.*?[\'"]{{lib}}[\'"]',
-        ],
-        Language.CPP: [
-            # #include <lib>  /  #include <lib/header.h>  /  #include "lib..."
-            r'^\s*#\s*include\s*[<"]{lib}[/\.>"]',
-        ],
-        Language.C: [
-            r'^\s*#\s*include\s*[<"]{lib}[/\.>"]',
-        ],
-    }
-
-    @property
-    def name(self):
-        return "forbidden_import"
-
-    @property
-    def description(self):
-        return t("io.forbidden_import.description")
-
-    @property
-    def required_file(self):
-        """Returns the name of the file required for this test, or None if not applicable."""
-        return None
-
-    @property
-    def parameter_description(self):
-        return [
-            ParamDescription(
-                "forbidden_imports",
-                t("io.forbidden_import.params.libraries"),
-                "list of strings"
-            ),
-            ParamDescription(
-                "submission_language",
-                t("io.forbidden_import.params.language"),
-                "string or Language enum"
-            ),
-        ]
-
-    # ------------------------------------------------------------------
-    # Helpers
-    # ------------------------------------------------------------------
-
-    def _build_patterns(self, library: str, language: Language) -> List[re.Pattern]:
-        """Return compiled regex patterns for detecting *library* in *language*."""
-        templates = self.IMPORT_PATTERNS.get(language, [])
-        compiled: List[re.Pattern] = []
-        for tmpl in templates:
-            # Support both {lib} and {{lib}} placeholders (Node patterns use
-            # double-braces to survive the first .format call on the class body).
-            raw = tmpl.replace('{{lib}}', library).replace('{lib}', re.escape(library))
-            compiled.append(re.compile(raw, re.MULTILINE))
-        return compiled
-
-    def _scan_file(self, content: str, forbidden: List[str],
-                   language: Language) -> List[str]:
-        """
-        Scan *content* for any forbidden imports.
-
-        Returns a list of human-readable violation strings.
-        """
-        violations: List[str] = []
-        for lib in forbidden:
-            patterns = self._build_patterns(lib, language)
-            for pattern in patterns:
-                match = pattern.search(content)
-                if match:
-                    violations.append(lib)
-                    break  # one match per library is enough
-        return violations
-
-    @staticmethod
-    def _resolve_language(submission_language=None) -> Optional[Language]:
-        """Resolve a raw language value into a Language enum member."""
-        if submission_language is None:
-            return None
-        if isinstance(submission_language, Language):
-            return submission_language
-        # Accept string values like "python", "java", etc.
-        for lang in Language:
-            if lang.value == str(submission_language).lower():
-                return lang
-        return None
-
-    # ------------------------------------------------------------------
-    # Execute
-    # ------------------------------------------------------------------
-
-    def execute(self, files: Optional[List[SubmissionFile]], sandbox: Optional[SandboxContainer],
-                *args, forbidden_imports: List[str] = None,
-                submission_language=None, **kwargs) -> TestResult:
-        """
-        Scan every submission file for forbidden imports.
-
-        Returns score 100 if no forbidden imports are found, 0 otherwise.
-        """
-        locale = kwargs.get("locale")
-        if not forbidden_imports:
-            return TestResult(
-                test_name=self.name,
-                score=100.0,
-                report=t("io.forbidden_import.report.no_imports", locale=locale)
-            )
-
-        language = self._resolve_language(submission_language)
-
-        if language is None:
-            return TestResult(
-                test_name=self.name,
-                score=0.0,
-                report=t("io.forbidden_import.report.no_lang", locale=locale)
-            )
-
-        if not files:
-            return TestResult(
-                test_name=self.name,
-                score=100.0,
-                report=t("io.forbidden_import.report.no_files", locale=locale)
-            )
-
-        all_violations: List[str] = []
-        for submission_file in files:
-            found = self._scan_file(
-                submission_file.content, forbidden_imports, language
-            )
-            for lib in found:
-                all_violations.append(
-                    t("io.forbidden_import.report.violation", locale=locale, lib=lib, file=submission_file.filename)
-                )
-
-        if all_violations:
-            details = "\n".join(all_violations)
-            return TestResult(
-                test_name=self.name,
-                score=0.0,
-                report=t("io.forbidden_import.report.failure", locale=locale, details=details)
-            )
-
-        return TestResult(
-            test_name=self.name,
-            score=100.0,
-            report=t("io.forbidden_import.report.success", locale=locale)
-        )
-
-
-# ===============================================================
-# TestFunction for Forbidden Keyword/Construct Detection
-# ===============================================================
-
-class ForbiddenKeywordConfig(BaseModel):
-    """Configuration schema for ForbiddenKeywordTest."""
-    forbidden_keywords: List[str] = Field(default_factory=list)
-    custom_ast_grep_rules: List[Dict[str, Any]] = Field(default_factory=list)
-
-class ForbiddenKeywordTest(TestFunction):
-    """
-    Tests that a submission does NOT use any of the specified forbidden
-    keywords or language constructs.
-
-    Uses ast-grep for structural analysis, which accurately detects
-    constructs even when they are not at the start of a line and
-    correctly ignores comments and string literals.
-    """
-
-    # Mapping of high-level keywords to language-specific ast-grep rules.
-    PREDEFINED_RULES: Dict[Language, Dict[str, Dict[str, Any]]] = {
-        Language.PYTHON: {
-            "for_loop": {"kind": "for_statement"},
-            "while_loop": {"kind": "while_statement"},
-            "eval_call": {"pattern": "eval($$$)"},
-            "exec_call": {"pattern": "exec($$$)"},
-        },
-        Language.JAVA: {
-            "for_loop": {"kind": "for_statement"},
-            "while_loop": {"kind": "while_statement"},
-        },
-        Language.NODE: {
-            "for_loop": {"kind": "for_statement"},
-            "while_loop": {"kind": "while_statement"},
-            "eval_call": {"pattern": "eval($$$)"},
-        },
-        Language.CPP: {
-            "for_loop": {"kind": "for_statement"},
-            "while_loop": {"kind": "while_statement"},
-            "do_while_loop": {"kind": "do_statement"},
-        },
-        Language.C: {
-            "for_loop": {"kind": "for_statement"},
-            "while_loop": {"kind": "while_statement"},
-            "do_while_loop": {"kind": "do_statement"},
-        },
-    }
-
-    @property
-    def name(self):
-        return "forbidden_keyword"
-
-    @property
-    def description(self):
-        return t("io.forbidden_keyword.description")
-
-    @property
-    def parameter_description(self):
-        return [
-            ParamDescription("forbidden_keywords", t("io.forbidden_keyword.params.keywords"), "list of strings"),
-            ParamDescription("custom_ast_grep_rules", t("io.forbidden_keyword.params.custom_rules"), "list of dicts"),
-        ]
-
-    @property
-    def config_schema(self) -> Type[BaseModel]:
-        """Optional Pydantic model to validate test parameters during tree building."""
-        return ForbiddenKeywordConfig
-
-    def execute(self, files: Optional[List[SubmissionFile]], sandbox: Optional[SandboxContainer],
-                *args, forbidden_keywords: List[str] = None,
-                custom_ast_grep_rules: List[Dict[str, Any]] = None,
-                structural_analysis: Optional[StructuralAnalysisResult] = None,
-                submission_language: Optional[Language] = None,
-                **kwargs) -> TestResult:
-        """
-        Scan target files for forbidden keywords or structural patterns.
-        """
-        locale = kwargs.get("locale")
-        forbidden_keywords = forbidden_keywords or []
-        custom_ast_grep_rules = custom_ast_grep_rules or []
-
-        if not forbidden_keywords and not custom_ast_grep_rules:
-            return TestResult(
-                test_name=self.name,
-                score=100.0,
-                report=t("io.forbidden_keyword.report.no_rules", locale=locale)
-            )
-
-        if structural_analysis is None:
-            # Fallback if structural analysis was not executed or failed
-            return TestResult(
-                test_name=self.name,
-                score=0.0,
-                report=t("io.forbidden_keyword.report.no_analysis", locale=locale)
-            )
-
-        if submission_language is None:
-            return TestResult(
-                test_name=self.name,
-                score=0.0,
-                report=t("io.forbidden_keyword.report.no_lang", locale=locale)
-            )
-
-        # Consolidate rules to run
-        active_rules: List[Dict[str, Any]] = list(custom_ast_grep_rules)
-        lang_predefined = self.PREDEFINED_RULES.get(submission_language, {})
-        
-        for kw in forbidden_keywords:
-            if kw in lang_predefined:
-                active_rules.append(lang_predefined[kw])
-            else:
-                # If a keyword is not predefined for this language, we skip it
-                # but might want to log it or report it as an internal warning.
-                pass
-
-        if not active_rules:
-            return TestResult(
-                test_name=self.name,
-                score=100.0,
-                report=t("io.forbidden_keyword.report.success", locale=locale)
-            )
-
-        all_violations: List[str] = []
-        
-        # Only check files passed to this test
-        if not files:
-            return TestResult(
-                test_name=self.name,
-                score=100.0,
-                report=t("io.forbidden_keyword.report.no_files", locale=locale)
-            )
-
-        for sub_file in files:
-            root = structural_analysis.roots.get(sub_file.filename)
-            if root is None:
-                continue
-
-            for rule in active_rules:
-                matches = root.root().find_all(**rule)
-                if matches:
-                    # Provide some feedback about what was found
-                    rule_desc = rule.get("kind") or rule.get("pattern") or str(rule)
-                    all_violations.append(
-                        t("io.forbidden_keyword.report.violation", 
-                          locale=locale, 
-                          rule=rule_desc, 
-                          file=sub_file.filename)
-                    )
-
-        if all_violations:
-            details = "\n".join(all_violations)
-            return TestResult(
-                test_name=self.name,
-                score=0.0,
-                report=t("io.forbidden_keyword.report.failure", locale=locale, details=details)
-            )
-
-        return TestResult(
-            test_name=self.name,
-            score=100.0,
-            report=t("io.forbidden_keyword.report.success", locale=locale)
-        )
-
-
 class InputOutputTemplate(Template):
     """
     A template for command-line I/O assignments. It uses the SandboxExecutor
@@ -708,8 +370,6 @@ class InputOutputTemplate(Template):
             "expect_output": ExpectOutputTest(),
             "dont_fail": DontFailTest(),
             "expect_file_artifact": ExpectFileArtifactTest(),
-            "forbidden_import": ForbiddenImportTest(),
-            "forbidden_keyword": ForbiddenKeywordTest(),
         }
 
     @property
